refactor: QueryLog 생성자 및 예외 상태코드 개선
1. QueryLog 생성자에 answer 필드를 추가하여 구조를 개선한다.
2. validateQueryLogInvariants 메서드에서 answer 필드 null 체크 로직을 추가한다.
3. QueryLogExceptionStatus에 ANSWER_CANNOT_BE_NULL 및 LLM_API_ERROR 예외 상태코드를 추가한다.

Resolves #19
