+++ src/main/java/com/posicube/assignment/querylog/domain/QueryLog.java
@@ -1,7 +1,7 @@
-package com.posicube.assignment.querylog.domain.vo;
+package com.posicube.assignment.querylog.domain;
 
 import com.posicube.assignment.common.exception.BaseException;
-import com.posicube.assignment.querylog.domain.ModelType;
+import com.posicube.assignment.querylog.domain.vo.ModelType;
 import com.posicube.assignment.querylog.exception.QueryLogExceptionStatus;
 import com.posicube.assignment.users.domain.entity.Users;
 import jakarta.persistence.*;
@@ -38,6 +38,9 @@ public class QueryLog {
     private String content;
 
     @Column(nullable = false)
+    private String answer;
+
+    @Column(nullable = false)
     private Long usedTokens;
 
     @Column(nullable = false, precision = 19, scale = 4)
@@ -49,21 +52,22 @@ public class QueryLog {
     private static final int MAX_QUERY_LENGTH = 800;
     private static final BigDecimal THOUSAND = new BigDecimal("1000");
 
-    private QueryLog(Users user, String q, ModelType type, Long usedTokens) {
-        validateQueryLogInvariants(user, q, type, usedTokens);
+    private QueryLog(Users user, String q, ModelType type, String answer, Long usedTokens) {
+        validateQueryLogInvariants(user, q, type, answer, usedTokens);
         this.user = user;
         this.type = type;
         this.content = q.trim();
+        this.answer = answer;
         this.usedTokens = usedTokens;
         this.cost = calculateCost(this.usedTokens, type);
         createAt = LocalDateTime.now();
     }
 
-    public static QueryLog create(Users user, String q, ModelType type, Long usedTokens) {
-        return new QueryLog(user, q, type, usedTokens);
+    public static QueryLog create(Users user, String q, ModelType type, String answer, Long usedTokens) {
+        return new QueryLog(user, q, type, answer, usedTokens);
     }
 
-    private static void validateQueryLogInvariants(Users user, String q, ModelType type, Long usedTokens) {
+    private static void validateQueryLogInvariants(Users user, String q, ModelType type, String answer, Long usedTokens) {
         if (Objects.isNull(user)) {
             throw new BaseException(QueryLogExceptionStatus.USER_CANNOT_BE_NULL);
         }
@@ -74,6 +78,10 @@ public class QueryLog {
 
         if (Objects.isNull(type)) {
             throw new BaseException(QueryLogExceptionStatus.MODEL_TYPE_CANNOT_BE_NULL);
+        }
+
+        if (Objects.isNull(answer)) {
+            throw new BaseException(QueryLogExceptionStatus.ANSWER_CANNOT_BE_NULL);
         }
 
         if (Objects.isNull(usedTokens)) {
